py-producer/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
uWSGI==2.0.21
Werkzeug==2.2.3
